Moving gql file closer to the domain
