--- fbg-server/src/models/UserInput.model.ts
@@ -1,6 +1,0 @@
-import { InputType } from '@nestjs/graphql';
-
-@InputType()
-export class UserInput {
-  nickname: string;
-}
