+++ fbg-server/src/users/gql/User.gql.ts
@@ -5,7 +5,9 @@ export class User {
   @Field((type) => Int, { nullable: true })
   id?: number;
 
+  @Field((type) => String, { nullable: false })
   nickname: string;
 
+  @Field((type) => String, { nullable: true })
   jwtToken?: string;
 }
