+++ fbg-server/src/users/gql/UserInput.gql.ts
@@ -1,0 +1,7 @@
+import { Field, InputType } from '@nestjs/graphql';
+
+@InputType()
+export class UserInput {
+  @Field((type) => String)
+  nickname: string;
+}
